--- app/fee_constants.py
@@ -1,4 +1,0 @@
-MORTGAGE_REGISTRATION_FEE = 85
-LAND_TRANSFER_BASE_FEE = 108.40
-LAND_TRANSFER_RATE = 2.46  # per $1,000
-CONVEYANCING_FEE = 1500
