--- app/suburb.py
@@ -1,9 +1,0 @@
-class Suburb:
-
-    def __init__(self, name, postcode=0000, council_name="NullCouncil", population_in_2016=0, crimes_against_person_2016=0, crimes_against_person_2019=0):
-        self.name = name
-        self.postcode = postcode
-        self.council_name = council_name
-        self.population_in_2016 = population_in_2016
-        self.crimes_against_person_2016=crimes_against_person_2016
-        self.crimes_against_person_2019=crimes_against_person_2019
