+++ app/templates/listings/index.html
@@ -38,7 +38,6 @@
                                         </div>
                                     {% endif %}
                                 </div>
-                                <a class="btn btn-primary" href="{{ url_for('get_listing', property_id=listing['id']) }}">See more</a>
                             </div>
                         </div>
                     </div>
